Use external kustomize command
To better guard against compatibility issues with whatever
version of kustomize is included with the customer's
kubectl command, we have changed to use an external kustomize
binary.  This should allow the customer to better control what
version of kustomize they can use.

Updated the manifest files to be compatibility with latest
stable release of kustomize.  Also updated instructions.