Right-sized cpu requests based on perf from GKE in us-central
Based on load tests, adjusted cpu limits and the amount of logs
generate per pod.  We need to do some investigation later on
to understand why large requests seem to be penalized by the
network.  For now though, these numbers provide for predictable
load generation at the least.

Also turned off raw_timestamp in anticipation for broader
roll out of this change now that it is not necessary.